build!: Restructure tests in build process
Now all CI tests can be run with a single "dotnet test" from the solution directory.

* Removed the functional test from the solution file.
* Combined code coverage into a single config file.

This has the added benefit of pulling code coverage into a single project-wide view.

Change was needed because test projects are starting to proliferate, and it will only get worse as I divide out components and add new ones for SPA.

Note that this is POTENTIALLY BREAKING. I'm not 100% sure whether the build will pass in ADO yet. If it breaks, I'll keep at it. Also I only moved the dotnet pipeline over. Once I'm happy with how that's working, I can port the changes to the other. Hopefully this should NOT break the others.

For AB#1433: File reorganization